+++ Editor/Authoring/MultiplayApis/MultiplayApiFactory.cs
@@ -8,6 +8,8 @@ using Unity.Services.Multiplay.Authoring.Editor.AdminApis.Allocations.Http;
 using Unity.Services.Multiplay.Authoring.Editor.AdminApis.BuildConfigs.Apis.BuildConfigurations;
 using Unity.Services.Multiplay.Authoring.Editor.AdminApis.Builds.Apis.Builds;
 using Unity.Services.Multiplay.Authoring.Editor.AdminApis.Fleets.Apis.Fleets;
+using Unity.Services.Multiplay.Authoring.Editor.AdminApis.Servers;
+using Unity.Services.Multiplay.Authoring.Editor.AdminApis.Servers.Apis.Servers;
 using Unity.Services.Multiplay.Authoring.Editor.Shared.Clients;
 using UnityEditor;
 
@@ -15,7 +17,7 @@ using ICoreAccessTokens = Unity.Services.Core.Editor.IAccessTokens;
 
 namespace Unity.Services.Multiplay.Authoring.Editor.MultiplayApis
 {
-    class MultiplayApiFactory : IFleetApiFactory, IBuildsApiFactory, IBuildConfigApiFactory, IAllocationApiFactory
+    class MultiplayApiFactory : IFleetApiFactory, IBuildsApiFactory, IBuildConfigApiFactory, IAllocationApiFactory, IServersApiFactory
     {
         const string k_StgUrl = "https://staging.services.unity.com";
 
@@ -58,11 +60,18 @@ namespace Unity.Services.Multiplay.Authoring.Editor.MultiplayApis
         async Task<IAllocationApi> IAllocationApiFactory.Build()
         {
             var(config, basePath, headers) = await Authenticate();
-            //TODO pass the config (e.g. project and environments)
-            return new AllocationApi(
+            return new AllocationApi(config,
                 new AllocationsApiClient(
                     new HttpClient(),
                     new AdminApis.Allocations.Configuration(basePath, null, null, headers)));
+        }
+
+        async Task<IServersApi> IServersApiFactory.Build()
+        {
+            var(config, basePath, headers) = await Authenticate();
+            return new ServersApi(config, new ServersApiClient(
+                new AdminApis.Servers.Http.HttpClient(),
+                new Configuration(basePath, null, null, headers)));
         }
 
         async Task<(ApiConfig, string, IDictionary<string, string>)> Authenticate()