+++ Runtime/Authoring/Core/Deployment/IDeploymentFacade.cs
@@ -1,3 +1,4 @@
+using System;
 using System.Collections.Generic;
 using System.Threading;
 using System.Threading.Tasks;
@@ -11,12 +12,13 @@ namespace Unity.Services.Multiplay.Authoring.Core.Deployment
     interface IDeploymentFacade
     {
         Task BuildBinaryAsync(BuildItem buildItem, CancellationToken cancellationToken = default);
-        Task RevertToOriginalBuildTargetAsync(CancellationToken cancellationToken = default);
+        Task WarnBuildTargetChanged(CancellationToken cancellationToken = default);
         Task<BuildUploadResult> UploadBuildAsync(BuildItem buildItem, CancellationToken cancellationToken = default);
         Task<bool> SyncBuildAsync(BuildItem buildItem, BuildId buildId, CloudBucketId bucketId, CancellationToken cancellationToken = default);
         Task<BuildId> FindBuildAsync(BuildName name, CancellationToken cancellationToken = default);
         Task<BuildConfigurationId> DeployBuildConfigAsync(BuildConfigurationName name, BuildId buildId, MultiplayConfig.BuildConfigurationDefinition definition, CancellationToken cancellationToken = default);
         Task<BuildConfigurationId> FindBuildConfigAsync(BuildConfigurationName name, CancellationToken cancellationToken = default);
         Task<FleetId> DeployFleetAsync(FleetName name, IList<BuildConfigurationId> buildConfigs, MultiplayConfig.FleetDefinition definition, CancellationToken cancellationToken = default);
+        Task<AllocationInformation> CreateAndSyncTestAllocationAsync(FleetName fleetName, BuildConfigurationName buildConfigurationName, CancellationToken cancellationToken = default);
     }
 }