+++ Runtime/Authoring/Core/MultiplayApi/AllocationResult.cs
@@ -2,5 +2,5 @@ using System;
 
 namespace Unity.Services.Multiplay.Authoring.Core.MultiplayApi
 {
-    record AllocationResult(Guid Id);
+    record AllocationResult(Guid AllocationId, string Href);
 }