+++ Runtime/Authoring/Core/MultiplayApi/FleetInfo.cs
@@ -1,4 +1,5 @@
 using System;
+using System.Collections.Generic;
 
 namespace Unity.Services.Multiplay.Authoring.Core.MultiplayApi
 {
@@ -10,12 +11,14 @@ namespace Unity.Services.Multiplay.Authoring.Core.MultiplayApi
             Status fleetStatus,
             Guid osId,
             string osName,
+            List<FleetRegionInfo> regions,
             AllocationStatus allocationStatus = null)
         {
             FleetName = fleetName;
             FleetStatus = fleetStatus;
             OsId = osId;
             OSName = osName;
+            Regions = regions;
             Id = id;
             Allocation = allocationStatus;
         }
@@ -25,6 +28,7 @@ namespace Unity.Services.Multiplay.Authoring.Core.MultiplayApi
         public Status FleetStatus { get; }
         public Guid OsId { get; }
         public string OSName { get; }
+        public List<FleetRegionInfo> Regions { get; }
         public AllocationStatus Allocation { get; }
 
         public enum Status
@@ -35,5 +39,6 @@ namespace Unity.Services.Multiplay.Authoring.Core.MultiplayApi
         }
 
         public record AllocationStatus(int Total, int Allocated, int Available, int Online);
+        public record FleetRegionInfo(Guid Id, Guid RegionId, string Name);
     }
 }