+++ Runtime/Authoring/Core/MultiplayApi/IServersApi.cs
@@ -1,0 +1,11 @@
+using System;
+using System.Threading;
+using System.Threading.Tasks;
+
+namespace Unity.Services.Multiplay.Authoring.Core.MultiplayApi
+{
+    interface IServersApi
+    {
+        Task<bool> TriggerServerActionAsync(long serverId, ServerAction action, CancellationToken cancellationToken = default);
+    }
+}