+++ Runtime/Authoring/Core/MultiplayApi/ServerAction.cs
@@ -1,0 +1,14 @@
+using System;
+
+namespace Unity.Services.Multiplay.Authoring.Core.MultiplayApi
+{
+    enum ServerAction
+    {
+        Add = 1,
+        Delete = 2,
+        Restart = 3,
+        Shutdown = 4,
+        Start = 5,
+        Stop = 6
+    }
+}